Colilsion system impl with ecs

+++ BrawlKingTest.Unity/Assets/GameCore/ECS/Components/Interfaces/IHealthComponent.cs
@@ -1,6 +1,5 @@
 ﻿interface IHealthComponent
 {
-    void TakeDamage(int value);
-    void TakeHeal(int value);
-
+    void TakeDamage(float value);
+    void TakeHeal(float value); 
 }
